feat: update files to latest changes in project

+++ src/api/controllers/products.controller.ts
@@ -1,9 +1,13 @@
 import { apiConfig } from '../../config/apiConfig.ts';
-import { IProduct, IProductResponse, IProductsResponse } from '../../data/types/product.types.ts';
-import { IRequestOptions } from '../../data/types/api.types.ts';
+import {
+  IProduct,
+  IProductRequestParams,
+  IProductResponse,
+  IProductsResponse,
+  IRequestOptions,
+} from '../../data/types/index.ts';
 import { AxiosApiClient } from '../apiClients/axios.apiClient.ts';
 import { logStep } from '../../utils/reporter/decorators.ts';
-import { IProductRequestParams } from '../../data/types/requestParams.ts';
 import { convertRequestParams } from '../../utils/request.ts';
 
 class ProductsController {
@@ -18,14 +22,17 @@ class ProductsController {
         Authorization: token,
       },
       data: productData,
-      url: apiConfig.endpoints.Products,
-      baseURL: apiConfig.baseUrl,
+      url: apiConfig.ENDPOINTS.PRODUCTS,
+      baseURL: apiConfig.BASE_URL,
     };
+
     return await this.apiClient.send<IProductResponse>(options);
   }
 
   @logStep('Update Product via API')
   async update(productData: IProduct, productId: string, token: string) {
+    if (!productId) throw new Error('Product ID is required and cannot be null or undefined');
+
     const options: IRequestOptions = {
       method: 'put',
       headers: {
@@ -33,29 +40,32 @@ class ProductsController {
         Authorization: token,
       },
       data: productData,
-      url: apiConfig.endpoints['Get Product By Id'](productId),
-      baseURL: apiConfig.baseUrl,
+      url: apiConfig.ENDPOINTS.PRODUCT_BY_ID(productId),
+      baseURL: apiConfig.BASE_URL,
     };
+
     return await this.apiClient.send<IProductResponse>(options);
   }
 
   @logStep('Get Product by id via API')
   async get(productId: string, token: string) {
+    if (!productId) throw new Error('Product ID is required and cannot be null or undefined');
+
     const options: IRequestOptions = {
       method: 'get',
       headers: {
         'content-type': 'application/json',
         Authorization: token,
       },
-      // TODO prevent requst sending if productID is null or undefined
-      url: apiConfig.endpoints['Get Product By Id'](productId),
-      baseURL: apiConfig.baseUrl,
+      url: apiConfig.ENDPOINTS.PRODUCT_BY_ID(productId),
+      baseURL: apiConfig.BASE_URL,
     };
+
     return await this.apiClient.send<IProductResponse>(options);
   }
 
   @logStep('Get All Products via API')
-  async getAll(token: string, params?: IProductRequestParams) {
+  async getAllWithParams(token: string, params?: IProductRequestParams) {
     let urlParams = '';
     if (params) {
       urlParams = convertRequestParams(params as Record<string, string>);
@@ -66,23 +76,42 @@ class ProductsController {
         'content-type': 'application/json',
         Authorization: token,
       },
-      url: apiConfig.endpoints.Products + urlParams,
-      baseURL: apiConfig.baseUrl,
+      url: apiConfig.ENDPOINTS.PRODUCTS + urlParams,
+      baseURL: apiConfig.BASE_URL,
     };
+
+    return await this.apiClient.send<IProductsResponse>(options);
+  }
+
+  @logStep('Get All Products (All) via API')
+  async getAll(token: string) {
+    const options: IRequestOptions = {
+      method: 'get',
+      headers: {
+        'content-type': 'application/json',
+        Authorization: token,
+      },
+      url: apiConfig.ENDPOINTS.PRODUCTS_ALL,
+      baseURL: apiConfig.BASE_URL,
+    };
+
     return await this.apiClient.send<IProductsResponse>(options);
   }
 
   @logStep('Delete Product by id via API')
   async delete(productId: string, token: string) {
+    if (!productId) throw new Error('Product ID is required and cannot be null or undefined');
+
     const options: IRequestOptions = {
       method: 'delete',
       headers: {
         'content-type': 'application/json',
         Authorization: token,
       },
-      url: apiConfig.endpoints['Get Product By Id'](productId),
-      baseURL: apiConfig.baseUrl,
+      url: apiConfig.ENDPOINTS.PRODUCT_BY_ID(productId),
+      baseURL: apiConfig.BASE_URL,
     };
+
     return await this.apiClient.send(options);
   }
 }
